% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{abundance_plot}
\alias{abundance_plot}
\title{Abundance Plot}
\usage{
abundance_plot(
  phylo,
  x,
  taxlev = "Family",
  orderbar = "",
  facetby = "Description",
  ntaxa = 14,
  bar.params = list(),
  facet.params = list(),
  ...
)
}
\arguments{
\item{phylo}{Phyloseq object}

\item{x}{String Variable for x axis}

\item{taxlev}{String Taxonomic Level}

\item{orderbar}{String. Numeric Variable of sample_data. The Bars will be
ordered in respect to the values}

\item{facetby}{String Variable the plot should be faceted by}

\item{ntaxa}{Integer number of taxa which should be displayed}

\item{bar.params}{Additional parameters for the geom_bar plot}

\item{facet.params}{Additional parameters for facet_wrap}

\item{...}{}
}
\value{
ggplot
}
\description{
Abundance Plot
}
\examples{
data("mice_B6_N")
abundance_plot(mice_B6_N, "Microbiota")
}
