% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{adonis_plot}
\alias{adonis_plot}
\title{Adonis Plot}
\usage{
adonis_plot(adonis_res, x, value, ylab, xlab = "", scale_x_discrete_limits)
}
\arguments{
\item{adonis_res}{adonis object}

\item{x}{List of variables to map}

\item{value}{Result values to map}

\item{ylab}{label for y axis}

\item{xlab}{label for x axis}

\item{scale_x_discrete_limits}{}
}
\value{
ggplot
}
\description{
Barplot for vegan::adonis results.
}
\examples{
adonis_plot(adonis_res, ...)
}
