% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_abun.R
\name{aldex2_join_tax}
\alias{aldex2_join_tax}
\title{Join ALDEx2 results with taxonomic information}
\usage{
aldex2_join_tax(phylo, aldex2_res)
}
\arguments{
\item{phylo}{phyloseq object}

\item{aldex2_res}{aldex2 object}
}
\value{
tibble
}
\description{
Join ALDEx2 results with taxonomic information
}
