% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_abun.R
\name{aldex2_res_sig}
\alias{aldex2_res_sig}
\title{ALDEx2 Significant OTUs}
\usage{
aldex2_res_sig(phylo, aldex2_res, sig = 0.05, all_sig = TRUE)
}
\arguments{
\item{phylo}{phyloseq object}

\item{aldex2_res}{aldex2 object}

\item{sig}{single numeric. Significance threshold}

\item{all_sig}{identify which values are siginifcant in both t-test and
glm test}
}
\value{
data.frame
}
\description{
Format ALDEx2 results returning taxonomic
information for significant results
}
\examples{
data("mice_B6_N")
aldex2_res <- ALDEx2::aldex(mice_B6_N@otu_table,
as.character(sample_data(phylo)$Vendor))
aldex2_res_sig(mice_B6_N, aldex2_res)
}
