% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diver_fun.R
\name{alpha_div_calc}
\alias{alpha_div_calc}
\title{Alpha diversity}
\usage{
alpha_div_calc(
  phylo,
  min_count = 2000,
  measures = c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson",
    "BF_ratio")
)
}
\arguments{
\item{phylo}{reads a phyloseq object}

\item{min_count}{samples that do not reach this count number}

\item{measures}{character vector. Type of alpha diversity indexes. Possible
values: c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson",
"BF_ratio"). "BF_ratio" calculates the ratio of Bacteroidetes and Firmicutes.}
}
\description{
Calulates alpha diversity distances
}
\examples{
data("mice_B6_N")
alpha_diversity_plus_dt(mice_B6_N, 2000)
}
\keyword{Alpha}
\keyword{diversity..}
