% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diver_fun.R
\name{alpha_div_plots}
\alias{alpha_div_plots}
\title{Alpha diversity plotting}
\usage{
alpha_div_plots(
  phylo,
  x_axis,
  color_var,
  measures = c("Observed", "Simpson", "Shannon", "BF_ratio"),
  jitter_params = list(),
  boxplot_params = list(),
  ...
)
}
\arguments{
\item{phylo}{phyloseq object with alpha-diversity distances in the
sample data}

\item{x_axis}{Determines x variable in plots. Column name in sample data
from phyloseq object as character}

\item{color_var}{Variable in sample data from phyloseq object
for color aesthetics}

\item{measures}{Alpha-diversity distances. Atomic character vector with
column names of sample data from the phyloseq object. Normally calculated
with alpha_div_calc()}

\item{jitter_params}{Additional parameters for ggplot2 geom_jitter}

\item{boxplot_params}{Additional parameters for ggplot2 geom_boxplot}

\item{...}{More Paremeters}
}
\description{
Plots boxplots of alpha-diversity distances with jitter plot.
}
\examples{
data("mice_B6_N")
phylo <- phyloseq::subset_samples(mice_B6_N, Description == "IN")
phylo <- phyloseq::prune_taxa(phyloseq::taxa_sums(mice_B6_N) > 0, phylo)
phylo_a_dist <- alpha_div_calc(phylo, 2000)
alpha_div_plots(phylo_a_dist, "Vendor", "Vendor")
}
\keyword{alpha-diversity}
\keyword{plot}
