% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diver_fun.R
\name{bacfir_ratio_calc}
\alias{bacfir_ratio_calc}
\title{Calculate bacteroides to firmicutes ratio}
\usage{
bacfir_ratio_calc(physeq = phylo)
}
\arguments{
\item{phylo}{phyloseq object}
}
\value{
phyloseq object with additional variable in sample_data
}
\description{
Calculate bacteroides to firmicutes ratio
}
\examples{
data("mice_B6_N")
bacfir_ratio(mice_B6_N)
}
