% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{beta_dist_calc}
\alias{beta_dist_calc}
\title{Beta diversity distances inter/intra samples}
\usage{
beta_dist_calc(phylo, var, dist, sample_ID = "SampleID")
}
\arguments{
\item{phylo}{phyloseq object}

\item{var}{variable in sample data of the phyloseq object, which should be
compared}

\item{dist}{Filter samples that do not reach this count number}

\item{sample_ID}{Sample identifier column}
}
\value{
data.frame
}
\description{
This function calculate beta diversity distances inter/intra samples and
returns a data frame
}
\examples{
data("mice_B6_N")
get_beta_distances(mice_B6_N, var = "Genotype", dist = "unifrac")

}
\keyword{beta}
\keyword{diversity}
