% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{beta_dist_plot}
\alias{beta_dist_plot}
\title{Plot Beta diversity distances inter/intra samples}
\usage{
beta_dist_plot(
  beta_distance,
  x = "comparison",
  y = "value",
  jitter = FALSE,
  alpha = 0.5
)
}
\arguments{
\item{beta_distance}{data frame from beta_dist_calc}

\item{x}{x axis. Descriptive data.}

\item{y}{y axis. Value data.}

\item{jitter}{if TRUE adds addition geom_jitter}

\item{alpha}{only for jitter, regulates alpha value. Possible values from
0 to 1}
}
\description{
This function
}
\examples{
data("mice_B6_N")
beta_dist <- beta_dist_calc(mice_B6_N, var = "Genotype", dist = "unifrac")
get_beta_distances_plot(beta_dist)
}
\keyword{Beta}
\keyword{diversity}
\keyword{plot}
