% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{beta_ordination}
\alias{beta_ordination}
\title{Plot beta ordination function}
\usage{
beta_ordination(
  phylo,
  formula = "",
  color,
  distance_method = "bray",
  ordination_type = "NMDS",
  geom_star = TRUE,
  permutations = 999,
  parallel = 4,
  shape = NULL,
  tagtype = "A",
  verbose = TRUE,
  layout = c(area(1, 1), area(2, 1, 5))
)
}
\arguments{
\item{phylo}{Phyloseq object}

\item{formula}{stats for adonis i.e. c("Genotype + Experiment + Cage.ID")}

\item{color}{metadata category color}

\item{distance_method}{Similarity or distance measurement i.e. "bray",
"jaccard", "uunifrac", "wunifrac"}

\item{ordination_type}{Type of ordination that should be performed}

\item{geom_star}{boolean. If TRUE create a star plot by drawing segments from
group centroid to each points.}

\item{permutations}{a list of control values for the permutations as
returned by the function how, or the number of permutations required,
or a permutation matrix where each row gives the permuted indices.}

\item{parallel}{Parallel computations}

\item{shape}{metadata category shape}

\item{tagtype}{Label type of the plots. i.e. "A" "1" "I".}

\item{verbose}{If TRUE calculation results will be printed. If FALSE not.}

\item{layout}{List of patch_area objects. Control the plot layout.}
}
\description{
This function performs ordination plots and adonis test
}
\examples{
data("mice_B6_N")
beta_ordination(mice_B6_N, mycolor = "Genotype", myshape = "Experiment",
myformula = c("Genotype + Experiment + Cage.ID"))
}
\keyword{Beta}
\keyword{diversity}
