% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_rel_abund}
\alias{calc_rel_abund}
\title{Calculate relative abundance}
\usage{
calc_rel_abund(phylo, n = 1)
}
\arguments{
\item{phylo}{Phyloseq object}
}
\value{
phyloseq object
}
\description{
Calculate relative abundance
}
\examples{
data("mice_B6_N")
calc_rel_abund(mice_B6_N)
}
