% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_randomForest}
\alias{classify_randomForest}
\title{Random Forest Model}
\usage{
classify_randomForest(
  phylo,
  classify_by,
  minlib = 1500,
  prunescale = 0.01,
  seed = 2,
  ntree = 500
)
}
\arguments{
\item{phylo}{phyloseq object}

\item{classify_by}{variable with which the data should be classified by}

\item{minlib}{normalizing to minimum reads}

\item{prunescale}{minimum abundance in percantage an OTU must have to be considered}

\item{seed}{set seed for reproducable research}

\item{ntree}{number of trees to grow}
}
\value{
randomForest object
}
\description{
Creates a classification model based on Random Forest
}
\examples{
data("mice_B6_N")
classify_randomForest(mice_B6_N, "treatment_group")
}
