% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_abun.R
\name{comp_diff_abund_analysis}
\alias{comp_diff_abund_analysis}
\title{Compare Differential Abundance Analysis}
\usage{
comp_diff_abund_analysis(x, y, type, by = "OTU")
}
\arguments{
\item{x}{data.frame with significant observations as rownames}

\item{y}{data.frame with significant observations as rownames}

\item{by}{identifier in both tables}
}
\value{
list
}
\description{
Returns matching or missmatching rows from the
Differential Abundance Analysis of different methods
}
\examples{
data("mice_B6_N")
deseq2_res <- deseq2_analysis(mice_B6_N, "Mircobiota")
aldex2_res <- ALDEx2::aldex(mice_B6_N@otu_table,
as.character(sample_data(phylo)$Vendor))
comp_diff_abund_analysis(deseq2_res, aldex2_res, type = "match")
}
