% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{Read 16S amplicon biom file into phyloseq object}
\usage{
data_load(otu, map, tree, min_otu = 0.2, min_sample = 999, verbose = FALSE)
}
\arguments{
\item{otu}{character string. Otu_table file path. In biom file oldformat
from qiime 1.8.}

\item{map}{character string. Mapping file path. In biom file oldformat from
qiime 1.8.}

\item{tree}{character string. tree file path. In nexus format from qiime 1.8.}

\item{min_otu}{numeric. Filter OTUs that do not reach this count number}

\item{min_sample}{numeric. Filter samples that do not reach this count number}

\item{verbose}{boolean. If TRUE cat()s step progress}
}
\description{
This function allows you to load the data recived from amplicon sequencing
from MIKI-lab.
}
\examples{
phylo <- MIKIbiomeR::data_load(otu = otu_path, map = map_path,
tree = tree_path)
}
\keyword{amplicon}
\keyword{load}
