% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_abun.R
\name{deseq2_analysis}
\alias{deseq2_analysis}
\title{Differential expression analysis}
\usage{
deseq2_analysis(
  phylo,
  var,
  alpha = 0.01,
  test = "Wald",
  fitType = "parametric"
)
}
\arguments{
\item{phylo}{Phyloseq Object with OTU table and sample data}

\item{var}{Variable of sample data to analyze}

\item{test}{significance test used in DESeq2. Either "Wald" or "LRT"}

\item{fitType}{type of fitting of dispersions to the mean intensity
in DESeq2. Either "parametric", "local" or "mean"}
}
\value{
DESeqDataSet object including data set and test results
}
\description{
Test for differential expression based on
negative binomal distribution
}
\examples{
data("mice_B6_N")
desq2_analysis(mice_B6_N, "Vendor")
}
