% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_abun.R
\name{deseq2_plot}
\alias{deseq2_plot}
\title{Plot DESeq2 analysis}
\usage{
deseq2_plot(
  deseq2_result,
  title,
  label_taxlev,
  top_hits = 10,
  fdr = 0.05,
  fc = 2,
  size = 2,
  alpha = 0.3
)
}
\arguments{
\item{deseq2_result}{deseq2 object}

\item{title}{Title of the plot}

\item{label_taxlev}{Taxonomic label for the top hits}

\item{top_hits}{Top hits which should be labled}

\item{fdr}{single numeric. false discovery rate}

\item{fc}{single numeric. fold change}

\item{size}{single numeric}

\item{alpha}{single numeric.}
}
\value{
ggplot object
}
\description{
Plots results from DESeq2 analysis
}
\examples{
data("mice_B6_N")
deseq2_res <- deseq2_analysis(mice_B6_N, "Mircobiota")
deseq2_plot(deseq2_res, "Deseq", "Family")
}
