% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity_fun.R
\name{ordination_plot}
\alias{ordination_plot}
\title{Plot Ordination}
\usage{
ordination_plot(
  phylo,
  ordination,
  type = "samples",
  color = NULL,
  shape = NULL,
  geom_star = TRUE
)
}
\arguments{
\item{phylo}{phyloseq object}

\item{ordination}{ordination object}

\item{type}{Plot type. Supported options are
c("samples", "sites", "species", "taxa", "biplot", "split", "scree")}

\item{color}{Map colors based on character string
of a sample variable or taxonomic rank}

\item{shape}{Map shapes based on character string of a sample variable
or taxonomic rank}

\item{geom_star}{boolean. If TRUE create a star plot by drawing segments from
group centroid to each points.}
}
\value{
ggplot
}
\description{
Plot Ordination
}
\examples{
data("mice_B6_N")
ordination <- phyloseq::ordinate(mice_B6_N, "NMDS", "bray")
ordination_plot(mice_B6_N, ordination)
}
