% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{phyloseq_to_tidy}
\alias{phyloseq_to_tidy}
\title{Export Phyloseq as tidy data frame}
\usage{
phyloseq_to_tidy(phylo)
}
\arguments{
\item{phylo}{Phyloseq object

It is slow, so only use when necessary}
}
\value{
tidy data frame
}
\description{
Export Phyloseq as tidy data frame
}
\examples{
data("mice_B6_N")
phyloseq_to_tidy(mice_B6_N)
}
