% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{plot_randomForest_results}
\alias{plot_randomForest_results}
\title{Plot Random Forest Results}
\usage{
plot_randomForest_results(randomForest_obj, phylo, classified_by, top = 20)
}
\arguments{
\item{randomForest_obj}{randomForest object}

\item{top}{number of top OTUs, which should be plotted}
}
\value{
plot
}
\description{
Plot Random Forest Results
}
\examples{
data("mice_B6_N")
prediction_model <- classify_randomForest(mice_B6_N, "treatment_group")
plot_randomForest_results(prediction_model, 10)
}
