% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tax_subset}
\alias{tax_subset}
\title{Summarises Phyloseq object to taxlevel and highest n abundance}
\usage{
tax_subset(phylo, top_n, taxlev)
}
\arguments{
\item{phylo}{phyloseq object}

\item{top_n}{single numeric, prunes taxa with highest n abundance}

\item{taxlev}{character. Taxonomic level.}
}
\value{
phyloseq object
}
\description{
Summarises Phyloseq object to taxlevel and highest n abundance
}
